Elaborators check for generator results in both cached and from prior elabs. Built-in pdk modules renamed 'pdk'
