+++ pdks/Asap7/asap7/__init__.py
@@ -1,11 +1,13 @@
-"""
-# Hdl21 ASAP7 PDK 
-"""
-
+from typing import Optional
 from hdl21.pdk import register
 
 # Grab our primary PDK-definition module
-from . import asap7
+from . import pdk
+from .pdk import *
+
+# The optional external-data installation.
+# Set by an instantiator of `Install`, if available.
+install: Optional[Install] = None
 
 # And register as a PDK module
-register(asap7)
+register(pdk)
