+++ pdks/Sky130/sky130/__init__.py
@@ -2,13 +2,12 @@ from typing import Optional
 from hdl21.pdk import register
 
 # Grab our primary PDK-definition module
-from . import sky130
-from .sky130 import modules, Sky130MosParams, Sky130Walker, Install
-
+from . import pdk
+from .pdk import *
 
 # The optional external-data installation.
 # Set by an instantiator of `Install`, if available.
 install: Optional[Install] = None
 
 # And register as a PDK module
-register(sky130)
+register(pdk)
